requirements.txt
click==8.1.3
Flask==2.2.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.2.2
requests==2.22.0
python-dotenv==0.21.0
Flask-SQLAlchemy==3.0.2
greenlet==2.0.0
importlib-metadata==5.0.0
psycopg2-binary==2.9.5
SQLAlchemy==1.4.42
zipp==3.10.0
flask-login==0.6.2